sm8750-common: vibrator: translate compose primitives to effects
The primitive 8 (LOW_TICK) is widely used through the compose API in
system UI components like the volume slider, the brightness slider, and
the task switcher.

Compose primitives always set the 12th bit of the effect id to 1, so
the current getEffectStream cannot find a matching effect from the
loaded effect files. The function then falls back to the predefined
primitives in effects.cpp, which is suboptimal and does not have a
config for primitive 8.

However, some effects from the binary files can be used as replacements
for compose primitives. For example, effect 305 can substitute primitive
8 because of the following reasons.

1. Effect 305 is short, so the compose API can use it to create longer
   consecutive ticks.
2. Effect 305 has a relative low amplitude, which works well for subtle
   haptic feedbacks in scenarios like dragging sliders and switching tasks
   in the task switcher.

This commit introduces a function to translate some compose primitives
to existing effects. As of writing, I only notice primitive 8 being used
by the android framework. Other vibration effects like the click and
tick are still played via the aidl v1 interface.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: NoCache-69 <chethangowdasj86@gmail.com>